Assignment 3-Part 1/requirements.txt
alembic==1.5.8
altair==4.1.0
apache-airflow==2.0.1
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.2
apispec==3.3.2
argcomplete==1.12.2
argon2-cffi==20.1.0
asn1crypto==1.4.0
astor==0.8.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
azure-common==1.1.27
azure-core==1.12.0
azure-storage-blob==12.8.0
Babel==2.9.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
boto3==1.17.40
botocore==1.20.40
cached-property==1.5.2
cachetools==4.2.1
cattrs==1.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
ConfigArgParse==1.4
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
diagrams==0.19.1
dill==0.3.3
dnspython==2.1.0
docutils==0.16
email-validator==1.1.2
entrypoints==0.3
fastapi==0.63.0
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-BasicAuth==0.2.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flat-table==1.1.1
gevent==21.1.2
geventhttpclient==1.4.4
gitdb==4.0.7
GitPython==3.1.14
graphviz==0.16
greenlet==1.0.0
gunicorn==19.10.0
h11==0.12.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
locust==1.4.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
mistune==0.8.4
msgpack==1.0.2
msrest==0.6.21
natsort==7.1.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
oscrypto==1.2.1
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pendulum==2.1.2
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prison==0.1.3
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
py==1.10.0
pyarrow==3.0.0
pycparser==2.20
pycryptodomex==3.10.1
pydantic==1.8.1
pydeck==0.6.1
Pygments==2.8.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pysnowflake==0.1.3
pytest==6.2.2
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rich==9.2.0
s3transfer==0.3.6
Send2Trash==1.5.0
setproctitle==1.2.2
six==1.15.0
smmap==4.0.0
snowflake-connector-python==2.4.1
snowflake-sqlalchemy==1.2.4
SQLAlchemy==1.4.3
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
starlette==0.13.6
streamlit==0.79.0
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.26.4
uvicorn==0.13.4
validators==0.18.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
